This is a summary...Edited the Read Me
So this is the more in depth description….

Edited the read me file to include my first words on here. Then I click
commit to save changes. Next I click on sync at the top right to sync
the changes to the web. This will save to my fork.

After that I can then send a pull request to the main project to accept
my new edit.